UI:filter feature works in join room
